+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Not released yet
 
+### Features
+
+* Add `encrypt_with_password()` and `decrypt_with_password()` functions to
+  encrypt and decrypt data
+
 ### Fixes
 
 * Add more missing vendor classes into stubs